requirements.txt
click==8.1.7
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.26.3
opencv-python==4.9.0.80
OPi.GPIO==0.5.2
pillow==10.3.0
